requirements.txt
discord.py
watchdog
python-dotenv
aiosqlite
aiohttp
psutil